+++ components/src/lib/jschema/properties/CollapsibleProperty.svelte
@@ -6,6 +6,7 @@
 	 * @typedef {Object} Props
 	 * @property {import("../../types/form").CollapsibleFormElement} formElement
 	 * @property {null|(() => void)} remove function passed by the parent that removes this element
+	 * @property {null|((oldKey: string, newKey: string) => void)} renameKey function passed by the parent that renames a key of this element
 	 * @property {boolean} editable
 	 * @property {null|(() => void)} [reset] function passed by the parent that resets this element to its default value (used only on top-level objects)
 	 * @property {null|(() => void)} [init] - Function passed by the parent that initializes a nullable element
@@ -18,6 +19,7 @@
 	let {
 		formElement = $bindable(),
 		remove,
+		renameKey,
 		editable,
 		children,
 		reset = null,
@@ -27,13 +29,14 @@
 	} = $props();
 
 	/** @type {boolean} */
-	let collapsed = $derived(formElement.collapsed);
+	let collapsed = $state(false);
 	/** @type {boolean} */
 	let hasErrors = $state(false);
 	formElement.hasErrors.subscribe((v) => (hasErrors = v));
 	/** @type {string[]} */
 	let errors = $state([]);
 	formElement.errors.subscribe((v) => (errors = v));
+	formElement.collapsed?.subscribe((c) => (collapsed = c));
 
 	let isNull = $state(false);
 
@@ -51,8 +54,11 @@
 			// prevent collapse when we are clicking on a property description
 			return;
 		}
-		collapsed = !collapsed;
-		formElement.collapsed = collapsed;
+		if (formElement.collapsed) {
+			formElement.collapsed.update((oldValue) => !oldValue);
+		} else {
+			collapsed = !collapsed;
+		}
 	}
 
 	/**
@@ -78,8 +84,7 @@
 		} else if (!value && 'setToNull' in formElement) {
 			formElement.setToNull();
 		}
-		formElement.collapsed = !value;
-		collapsed = formElement.collapsed;
+		formElement.collapsed.update(() => !value);
 	}
 
 	const showResetButton = $derived(
@@ -106,7 +111,7 @@
 					</button>
 					<div class="collapsible-label">
 						<div class="flex-fill">
-							<PropertyLabel {formElement} {editable} {remove} tag="span" />
+							<PropertyLabel {formElement} {editable} {remove} {renameKey} tag="span" />
 						</div>
 					</div>
 					<div class="collapsible-prop-actions">
@@ -136,7 +141,7 @@
 					class:collapse={collapsed}
 					class:show={!collapsed}
 				>
-					<div class="accordion-body p-1">
+					<div class="accordion-body p-0">
 						{#if showErrors}
 							{#each errors as error, index (index)}
 								<div class="alert alert-danger mb-1 py-1 px-2">{error}</div>
@@ -182,9 +187,15 @@
 		bottom: 0;
 		right: 0;
 		left: 0;
-		padding: 16px;
+		padding: 0 16px;
 		z-index: 250;
 		pointer-events: none;
+	}
+
+	.collapsible-label .flex-fill {
+		display: flex;
+		align-items: center;
+		height: 100%;
 	}
 
 	:global(.collapsible-label .property-description) {
@@ -197,5 +208,25 @@
 	.collapsible-prop-actions .form-switch {
 		margin-top: 16px;
 		margin-right: 8px;
+	}
+
+	:global(.collapsible-prop-header .btn-remove-property) {
+		margin-left: -6px;
+	}
+
+	:global(.collapsible-prop-header button),
+	:global(.collapsible-prop-header input) {
+		pointer-events: auto;
+	}
+
+	:global(.collapsible-prop-header .edit-key-input-wrapper) {
+		width: calc(100% - 120px);
+		display: inline-block;
+		background-color: white;
+		border-radius: var(--bs-border-radius);
+	}
+
+	:global(.collapsible-prop-header .invalid-feedback) {
+		display: none;
 	}
 </style>