+++ src/lib/components/v2/tasks/AddSingleTask.svelte
@@ -356,7 +356,7 @@
 	}
 </script>
 
-<StandardDismissableAlert message={taskSuccessMessage} />
+<StandardDismissableAlert bind:message={taskSuccessMessage} />
 
 <form
 	onsubmit={(e) => {